+++ src/react-apps/applications/dashboard/__tests__/components/base/serviceOverview.test.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { Provider } from 'react-redux';
 import * as renderer from 'react-test-renderer';
 import configureStore from 'redux-mock-store';
-import { ServicesOverview } from '../../../src/services/servicesOverview';
+import { ServicesOverview } from '../../../src/dashboardServices/serviceOverview/servicesOverview';
 
 describe('>>> components/base/serviceOverview.tsx --- Snapshot', () => {
   let mockStore: any;
