+++ src/react-apps/applications/dashboard/src/sagas/index.ts
@@ -2,8 +2,8 @@ import { SagaIterator, Task } from 'redux-saga';
 import { fork } from 'redux-saga/effects';
 import { sagaMiddleware } from '../store';
 
+import dashboardSaga from '../dashboardServices/dashboardSagas';
 import languageSagas from '../fetchLanguage/languageSagas';
-import dashboardSaga from '../services/dashboardSagas';
 
 function* root(): SagaIterator {
   yield fork(languageSagas);
